+++ web/src/features/trace-graph-view/utils/isLanggraphTrace.ts
